packages/irisml-tasks-training/irisml_tasks_training-0.0.132.tar.gz/irisml_tasks_training-0.0.132/test/test_evaluate_accuracy.py
import unittest
import torch
from irisml.tasks.evaluate_accuracy import Task


class TestEvaluateAccuracy(unittest.TestCase):
    def test_simple(self):
        predictions = torch.Tensor([[0.1, 0.2, 0.1], [0.9, 0.2, 0.2], [0.4, 0.4, 0.8]])
        targets = torch.Tensor([1, 0, 2])
        outputs = Task(Task.Config()).execute(Task.Inputs(predictions, targets))

        self.assertEqual(outputs.accuracy, 1.0)

        targets = torch.Tensor([2, 1, 1])
        outputs = Task(Task.Config()).execute(Task.Inputs(predictions, targets))
        self.assertEqual(outputs.accuracy, 0.0)

    def test_2d_targets(self):
        predictions = torch.Tensor([[0.1, 0.2, 0.1], [0.9, 0.2, 0.2], [0.4, 0.4, 0.8]])
        targets = torch.Tensor([[1], [0], [2]])
        outputs = Task(Task.Config()).execute(Task.Inputs(predictions, targets))

        self.assertEqual(outputs.accuracy, 1.0)

    def test_1d_predictions(self):
        predictions = torch.Tensor([1, 0, 2])
        targets = torch.Tensor([1, 0, 2])
        outputs = Task(Task.Config()).execute(Task.Inputs(predictions, targets))

        self.assertEqual(outputs.accuracy, 1.0)

    def test_1d_predictions_2d_targets(self):
        predictions = torch.Tensor([1, 0, 2])
        targets = torch.Tensor([[1], [0], [2]])
        outputs = Task(Task.Config()).execute(Task.Inputs(predictions, targets))

        self.assertEqual(outputs.accuracy, 1.0)
